Address rounding differences in bee_waggle_6
To construct this timeseries from the original data
we compute the sine and cosine of the head angle
of the bee. On different systems this can result in
slight differences in the data due to rounding. This
commit adds a known version of the dataset with a
different rounding than the original one and adds
a comment to the user when the checksum is not matched
exactly.